src/SensusJournal.Core/ValueObjects/Schedule.cs
using System.Text.RegularExpressions;

namespace SensusJournal.Core.ValueObjects;

public record Schedule
{
    private const string cronPattern = @"^(?#minute)(\*|(?:[0-9]|(?:[1-5][0-9]))(?:(?:\-[0-9]|\-(?:[1-5][0-9]))?|(?:\,(?:[0-9]|(?:[1-5][0-9])))*)) (?#hour)(\*|(?:[0-9]|1[0-9]|2[0-3])(?:(?:\-(?:[0-9]|1[0-9]|2[0-3]))?|(?:\,(?:[0-9]|1[0-9]|2[0-3]))*)) (?#day_of_month)(\*|(?:[1-9]|(?:[12][0-9])|3[01])(?:(?:\-(?:[1-9]|(?:[12][0-9])|3[01]))?|(?:\,(?:[1-9]|(?:[12][0-9])|3[01]))*)) (?#month)(\*|(?:[1-9]|1[012]|JAN|FEB|MAR|APR|MAY|JUN|JUL|AUG|SEP|OCT|NOV|DEC)(?:(?:\-(?:[1-9]|1[012]|JAN|FEB|MAR|APR|MAY|JUN|JUL|AUG|SEP|OCT|NOV|DEC))?|(?:\,(?:[1-9]|1[012]|JAN|FEB|MAR|APR|MAY|JUN|JUL|AUG|SEP|OCT|NOV|DEC))*)) (?#day_of_week)(\*|(?:[0-6]|SUN|MON|TUE|WED|THU|FRI|SAT)(?:(?:\-(?:[0-6]|SUN|MON|TUE|WED|THU|FRI|SAT))?|(?:\,(?:[0-6]|SUN|MON|TUE|WED|THU|FRI|SAT))*))$";
    private string _cronExpression;
    public string CronExpression
    {
        get { return _cronExpression; }
        init { _cronExpression = ValidateCronExpression(value); }
    }

    private Schedule() { }

    private string ValidateCronExpression(string value)
    {
        if (Regex.IsMatch(value, cronPattern))
        {
            return value;
        }

        throw new ArgumentException("The value is not a valid cron expression", nameof(value));
    }

    public static Schedule Create(string cronExpression)
    {
        return new Schedule { CronExpression = cronExpression };
    }
}